rename operational review to best practice
