+++ src/components/best-practice/esma-panel.tsx
