+++ src/components/best-practice/panel.ts
